Made the AIBS voxel connectome a specific usecase
Now CachedProjections is the default class that should in principle work with any data source. But one might have to write
wrappers for data lookup that are specific for a given data source.
I wrote AibsMcmProjections as such a wrapper specific for the AIBS data source.
Also wrote a dummy wrapper that returns random data for testing purposes.