Moved more functionality to the base class
The base class (CachedProjections) now can read annotations from nrrd files and hierarchy from json files.
This is because those two file types are well-documented and standardized. So we can expect implementations
for other data sources to be able to use them as well. If they don't want, they can still override it in
the derived classes. For example, in AibsMcmProjections, the hierarchy is built instead from the more
AIBS-specific structures file.