Integrated the new, higher flexibility with the diffusion script
The input has slightly changed. Now, one specifies with an argument the path to a json file that
configures which CachedProjections class to use. And which further inputs (such as the path to
the allen voxel model cache) to give it. This allows easy integration of new class implementations
that are more specific for different data sources.