Made parcellation_project logic independent of the data source
Now all logic that is likely to be common to all data sources (AIBS voxelized connectome, human functional
connectivity, etc.) is in a base class "ParcellationLevel". The source-specific logic is in derived classes.

The configuration file specifies with class to use.

Implemented a derived class for the AIBS voxelized connectome as a data source ("AibsLevel").