Removed all but the core files
Removed all code that is not used for a custom parcellation project.
Removed all flatmaps, parcellations, hierarchies, temporary files, etc. that accumulated during
the project. I propose that the file structure of the final project (including all the split/split/split ...)
be released instead as a data set on Zenodo.